Add files via upload
As humans and Ai will work together in the future, we must know the power of Ai and we should have proper knowledge of how to behave with Ai. So in this project, I have added an emotional module in which Friday(Ai assistance) will be angry when you will not behave properly with her and she will happy when you apologize to her.